Record verified DSV4-Flash tensor-map facts; correct the layer-ratio reading
Three-way mechanical verification (tools/verify_dsv4_tensor_map.py): all
1304 expected tensors present in the GGUF with 0 shape mismatches against
llama.cpp deepseek4.cpp create_tensor declarations, and the reference
model.py parameter shapes agree in the same direction. Establishes the
dims convention (GGUF dims == GGML ne, dims[0] contiguous; torch shape is
reversed) and the real indexer tensor names.

Corrects an earlier misreading: compress_ratios has 46 entries = 43 main
layers + 3 MTP stages; main-layer pattern is [0,0]+[4,128]x20+[4] (21
ratio-4 layers incl. 40/42, 20 ratio-128 layers incl. 41), NOT zeros at
40-42 -- the file carries compressor/indexer weights on those layers,
which settles it. KV math and concurrency estimates updated accordingly
(256K/slot ~0.92 GiB, not 0.8). Also records IQ2_XS signed-zero semantics
and marks the numpy dequant as not yet bit-exact (pure-Python reference
remains the only bit-exact oracle until it passes the golden gate).

Constraint: facts in notes must be measured values with reproducible commands
Rejected: keep the 19/19 layer split from the first read | contradicted by tensor presence
Confidence: high
Scope-risk: narrow (notes + offline verification tool)
Directive: rerun tools/verify_dsv4_tensor_map.py if the GGUF or llama.cpp checkout changes
Tested: 1304/1304 tensors, 0 mismatches, 0 extras on the real file
Not-tested: runtime-side consumption of these layouts (Phase 2+)

+++ loader/gguf_dequant.py
@@ -67,9 +67,9 @@ def dequantize_iq2_xs_block(block: bytes) -> list[float]:
             delta = db0 if group < 2 else db1
             base = 32 * ib32 + 8 * group
             for j in range(8):
-                magnitude = (grid >> (8 * j)) & 0xFF
+                magnitude = float((grid >> (8 * j)) & 0xFF)
                 if signs & KMASK_IQ2XS[j]:
-                    magnitude = -magnitude
+                    magnitude = -magnitude  # keeps -0.0, matching C float semantics
                 out[base + j] = _f32(delta * magnitude)
     return out
